requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
python-dotenv==1.0.0
requests==2.31.0
supabase==2.3.5
cloudinary==1.36.0
python-pptx==0.6.21
python-docx==0.8.11
PyPDF2==3.0.1
huggingface-hub==0.19.3
